+++ sdk/go/pulumi/resource_test.go
@@ -4,6 +4,7 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/blang/semver"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -17,6 +18,30 @@ type testProv struct {
 	ProviderResourceState
 	// equality identifier used for testing
 	foo string
+}
+
+type testResourcePackage struct {
+	version semver.Version
+}
+
+func (rp *testResourcePackage) ConstructProvider(name, typ string, urn string) (ProviderResource, error) {
+	return &testProv{}, nil
+}
+
+func (rp *testResourcePackage) Version() semver.Version {
+	return rp.version
+}
+
+type testResourceModule struct {
+	version semver.Version
+}
+
+func (rm *testResourceModule) Construct(name, typ string, urn string) (Resource, error) {
+	return &testRes{}, nil
+}
+
+func (rm *testResourceModule) Version() semver.Version {
+	return rm.version
 }
 
 func TestResourceOptionMergingParent(t *testing.T) {