+++ CHANGELOG.md
@@ -6,6 +6,15 @@ CHANGELOG
 - Allow `serializeFunction` to capture secrets.
   [#6013](https://github.com/pulumi/pulumi/pull/6013)
 
+- Parallelize bulk operations in NodeJS Automation API.
+  [#6022](https://github.com/pulumi/pulumi/pull/6022)
+
+## 2.16.2 (2020-12-23)
+
+- Fix a bug in the core engine that could cause previews to fail if a resource with changes had
+  unknown output property values.
+  [#6006](https://github.com/pulumi/pulumi/pull/6006)
+
 ## 2.16.1 (2020-12-22)
 
 - Fix a panic due to unsafe concurrent map access.